Extract components from Toolbar
