+++ src/scripts/App/Toolbar/Toolbar.tsx
@@ -1,4 +1,6 @@
 import React from "react";
+import NoneSelectedButtons from "./NoneSelectedButtons";
+import SomeSelectedButtons from "./SomeSelectedButtons";
 
 import "./Toolbar.css";
 
@@ -11,56 +13,6 @@ type Props = {
   onDelete: () => void;
 };
 
-type NotSelectedButtonsProps = {
-  linkMode: boolean;
-  onChangeLinkMode: () => void;
-  onCreateTask: () => void;
-};
-
-const NotSelectedButtons = ({
-  linkMode,
-  onChangeLinkMode,
-  onCreateTask,
-}: NotSelectedButtonsProps) => (
-  <>
-    <button
-      aria-label="Change Link Mode"
-      onClick={onChangeLinkMode}
-      className={`
-              Toolbar__button
-              Toolbar__button-change-link-mode
-              ${linkMode ? "Toolbar__button-change-link-mode--active" : ""}
-              iconButton
-            `}
-    />
-    <button
-      aria-label="Create Task"
-      onClick={onCreateTask}
-      className="Toolbar__button Toolbar__button-create-task iconButton"
-    />
-  </>
-);
-
-type SelectedButtonsProps = {
-  onComplete: () => void;
-  onDelete: () => void;
-};
-
-const SelectedButtons = ({ onComplete, onDelete }: SelectedButtonsProps) => (
-  <>
-    <button
-      aria-label="Delete Selected"
-      onClick={onDelete}
-      className="Toolbar__button Toolbar__button-delete-selected iconButton"
-    />
-    <button
-      aria-label="Complete Selected"
-      onClick={onComplete}
-      className="Toolbar__button Toolbar__button-complete-selected iconButton"
-    />
-  </>
-);
-
 const Toolbar = ({
   tasksSelected,
   linkMode,
@@ -71,9 +23,9 @@ const Toolbar = ({
 }: Props): JSX.Element => (
   <div className="Toolbar">
     {tasksSelected ? (
-      <SelectedButtons onComplete={onComplete} onDelete={onDelete} />
+      <SomeSelectedButtons onComplete={onComplete} onDelete={onDelete} />
     ) : (
-      <NotSelectedButtons
+      <NoneSelectedButtons
         linkMode={linkMode}
         onChangeLinkMode={onChangeLinkMode}
         onCreateTask={onCreateTask}
